Test ROBOT 1.10.0 release candidate
Temporary commit to allow building a development image to test the
release candidate of ROBOT 1.10.0.

The archive of that release candidate is not *publicly* available
(a GitHub account is required to follow the link to the archive), so for
the purpose of that test we assume that the robot.jar file has been
manually downloaded and is available in the top-level directory.