Include SubObjectPropertyOf axioms when reducing.
By default, ROBOT's `reduce` command does not take SubObjectPropertyOf
axioms into account, which leads it to leave in place some axioms that
are in fact redundant.

We pass the `--include-subproperties true` option to all calls to
`reduce` to change that behaviour.

This is configurable at the Makefile level (by overriding the
REDUCE_OPTIONS variable, either on the command line or in the custom
Makefile), _not_ at the ODK level (as we already do with RELAX_OPTIONS).