Use the new `--clean-obo` option.
Use ROBOT's new `--clean-obo` option when producing the OBO artefacts.

Since there is already a `obo_format_options` in the ODK configuration,
we reuse it, and set the default value to

  --clean-obo 'strict drop-untranslatable-axioms'

We default to dropping the untranslatable axioms because this has always
been the standard behaviour of the ODK (before this option existed, this
was done by forcibly grepping out the `owl-axioms` tag from the OBO
files).

I would prefer to default to `--clean-obo simple`, which has the
additional effect of forcibly removing GCI axioms. However this is a
consequential change, one that we should probably not force upon our
users.

We set the default value using some custom code, instead of statically
when declaring the `obo_format_options` field. This way, if a user has
already set this option to something in her configuration file, she will
still get the default behaviour of dropping the untranslatable axioms.
But we do _not_ touch the value set by the user if it already contains a
`--clean-obo` flag.