Ignore properties file from OBDA Protégé plugin.
The "OntopOBDA" plugin for Protégé automatically creates a
MYONT.properties file when Protégé is editing a MYONT.owl file. In many
cases that file is not wanted, so by default it should not be committed.

If users do want to commit that file (maybe because they changed the
configuration of the plugin, and want their settings saved), all they
have to do is to force Git to commit (`git add -f ...`).

closes #1220