Add GH action that updates OBO epm every week
The file is downloaded from bioregistry once per week;

I prefer this modus operandi over blindly updating what is on bioregistry as it gives me a chance to review it when chances happen upstream. The location where the file is stored is debatable - perhaps we can add a generic "resources" folder for such cases.