Allow changing the configuration schema transparently for users.
This commit puts in place the required infrastructure to allow us to
change the overall configuration schema (by renaming or moving options
around, e.g. taking several related options in the top-level 'project'
dictionary and grouping them into a dedicated subdictionary) while
preserving compatibility with existing configuration files.

This is done by inserting a `update_config_dict()` method in the
`Generator#load_config()` method, between the parsing of the YAML config
file and the construction of the `Project` object from the parsed
dictionary. The update method is intended to manipulate the parsed
dictionary as needed to transform it into a dictionary that matches the
new configuration schema. The core of the method uses a list of
(old,new) tuples where `old` is the path of an old configuration option,
and `new` is the path to the new option.

With that method in place, we can rename/move options at will. All we
have to do to ensure compatibility with existing configuration files is
to declare the proper tuples in the `update_config_dict()` method.

For example, say we want to rename the `use_base_merging` option in the
`import_group` section to `merge_base`: we can do that by adding the
following tuple:

  ('import_group.use_base_merging', 'import_group.merge_base')

Whenever the update method renames/moves keys, it emits a warning
to let users know that their configuration file uses an old syntax.

In addition, users can use the new subcommand `update-project` to create
an updated configuration file where all old config keys are
automatically replaced by their new equivalents.