Regroup all ROBOT-related options in a dedicated robot section.
This is proof-of-concept reorganisation of the configuration schema to
test the whole idea that we can make relatively big changes to the
schema while preserving compatibility with existing files.

We create a new subsection in the top-level `project` dictionary, called
`robot`, intended to host all ROBOT-related options.

That is, the following options and group, that are currently directly
under the top-level object:

  obo_format_options: ...
  robot_relax_options: ...
  robot_reduce_options: ...
  robot_plugins:
    plugins:
      - name: my_plugin
  robot_report:
    release_reports: True
    fail_on: ERROR
    ...

are now renamed and regrouped into a single `robot` section:

  robot:
    obo_format_options: ...
    relax_options: ...
    reduce_options: ...
    plugins:
      - name: my_plugin
    report:
      release_reports: True
      fail_on: ERROR
      ...