Extracts the version IRI of component mirrors (#1239)
* Extracts the version IRI of component mirrors

this patterns will be used soon for all imports, just want to make sure it is used for the component mirrors as well. I hope this wont crash when there is no version iri.

* Remove version IRI overwriting altogether in component mirror

This is so I can still see the version IRI at least when I want to in the component mirror.