Fix how seeding process deals with release artefacts.
There are several problems with the way the seeding process (1) copies
the release artefacts from the src/ontology directory to the top-level
release directory and (2) optionally commits those files.

For (1), the problem is that the seeding script manually copies the
files itself, and do so by assuming too many things about those files.
Notably, it assumes they all go straight to the top-level directory,
whereas some should actually go a subdirectory (for example, report
files, when they are released, go into $(RELEASEDIR)/reports, not into
$(RELEASEDIR) directly; so do subset files).

The problem with (2) is a consequence of (1): if not all release files
have been correctly copied to their correct expected location in the
top-level directory, then the seeding script will try to `git add` files
that may not actually exist.

We fix (1) by adding a new helper target, `copy_release_files`, to take
care of copying all release files to the release directory. It is called
internally by the `prepare_release` target, and also by the seeding
script. This ensures that the same logic to copy release files will
always be used.

Wr fix (2) by updating the `show_release_assets` target (whose sole
real purpose is to be used by the seeding script) to list the _actual_
release files: that is, the files that are in the release directory,
rather than the files under src/.