Remove code of conduct from the template.
An automatically generated code of conduct is of little value without
any indication that it is in fact endorsed by the project, instead of
having simply been left there because the project owner may not even
have realised the ODK generated it.

Project owner who wants a CoC should take the necessary steps to add one
themselves -- if they truly care about the values that a CoC is intended
to convey, they should not mind having to do that. If they say, "I want
a CoC but I don't want to do anything to get one", then they don't care
about a CoC, they care about being seen as having a CoC -- in other
words, it's virtue signalling, and it's worse than not having a CoC.

closes #1228