Fix the run test for the odk.py script.
When checking whether the odk.py script can be invoked, call it with the
`--help` option. Otherwise, it will exit with a non-zero code, which
will be interpreted as indicating that the script is not runnable
(which will in turn be interpreted as a sign that something went wrong
when building the ODK image).