Add manage_import_declarations option.
Add a new project-level option in the ODK configuration schema to
control whether the ODK should manage the import declarations in the
-edit file and the redirections in the XML catalog.

When enabled (it is by default), users should never have to manually
touch the import declarations or the redirections (though in the case of
the redirections, they _can_ insert additional redirections if they
want; they will not be removed as long as they are not in the
"odk-managed-catalog" group).

If disabled, then it will be the same as it was before: the ODK will not
touch the import declarations and the redirections, it will be up to the
users to ensure they are correctly updated when needed.

closes #1274