Do not inject import declarations/redirections at seed time.
When seeding a new repository, if the user already opted to manually
control the import declarations (manage_import_declarations: false),
then we do not generate the import declarations in the -edit file nor
the redirections in the XML catalog.

This basically means that the `manage_import_declarations` option is
honored both when a repository is updated and when it is seeded for the
first time.