Remove sparqlprog tools
The sparqlprog tools were last released in 2019, last modified in 2021,
and have been unusable in the ODK default configuration for most of the
1.5.x branch (from 1.5 to 1.5.3 included) without anyone ever
complaining about the issue.

Most likely no one is using those tools anymore.

We already announced, in the release notes for the 1.6 version, that
they were deprecated and slated for removal in 1.7, so we remove them
now.

closes #1254