Remove Ammonite
Ammonite is now replaced by Scala-CLI. In the 1.6.x branch, we kept it
alongside Scala-CLI to give Scala users some time to transition from one
interpreter to the next, but we made clear that Ammonite was slated for
removal in a future release. We remove it now, for the 1.7 release.

closes #1162