Add basic support for automatically generated bridge files.
Add a new type of products: bridge files. A bridge file is intended to
contain "bridging axioms" to allow proper integration of an ontology
into another.

Here, we allow the owner of an ontology project to define the bridges
they intend to produce. The defined bridges will be generated into the
`src/ontology/bridges` directory, and upon release will be copied to
`$(RELEASEDIR)/bridges`

We support (for now) two types of bridges: SSSOM-derived bridges and
custom-generated bridges.

For SSSOM-derived files, we provide a basic SSSOM/T-OWL ruleset with
some elementary rules (users will probably want to customize that
ruleset; they can also choose to use a custom ruleset from that start).

For custom-generated bridges, we simply generate a dummy Makefile rule,
that will have to be overriden in the custom Makefile (similarly to
other "custom" products).