Allow basic filtering when generating bridge files.
Add an option to the "BridgeProduct" configuration allowing to select
the mappings from which the bridge should be derived.

For example, given the following configuration:

```
bridge_group:
  products:
    - id: my-bridge
      sources:
        - my-mappings
      objects:
        - http://purl.obolibrary.org/obo/UBERON_
        - http://purl.obolibrary.org/obo/CL_
```

among all the mappings in the `my-mappings` set, only those that have an
object in the UBERON or CL namespaces will be used to produce the
bridge.

It is currently not planned to allow for more complex filtering. Users
should customise the ruleset (or use an entirely custom ruleset from the
start) if they do need more complex filtering.