Always use GitHub-CLI for public release to GitHub.
Remove the 'github_curl' and 'github_python' methods of pushing a
release to GitHub -- these should no longer be needed now that
GitHub-CLI is available.

The 'public_release' ODK option should now be set to the name of the
hosting service for releases. For now, the only supported option is
'github' (which, in effect, is exactly as before). The older values
'github_curl' and 'github_python' are silently transformed into
'github'.

To accomodate for that change:

(1) GitHub-CLI (`gh`) is moved from ODKFull to ODKLite;

(2) conversely, PyGithub (which was only used for the `github_python`
    method) is moved from ODKLite to ODKFull (it is not completely
    *removed*, since people may have custom workflows that depend on
    it).

closes #1296