Rely on odk-core to install ODKLite Python packages.
All the Python packages that are required to run ODK workflows are
dependencies of the odk-core package (with the `workflows` extra). This
means that, when building ODKLite, the only Python package we need to
explicitly install is odk-core.

We can thus remove the `requirements.txt.lite` file that is no longer
needed (we explicitly invoke `pip install odk-core[workflows]` instead),
and consequently rename `requirements.txt.full` into `requirements.txt`.