Re-organise the image around ODK-Core.
ODK-Core expects a certain layout for a "ODK environment". At the very
least, it expects a "resources" directory (which should be pointed to by
a "ODK_RESOURCES_DIR" variable, with the following contents:

* ODK_RESOURCES_DIR/obo.epm.json (the OBO extended prefix map);
* ODK_RESOURCES_DIR/robot/profile.txt (current version of the default
  ROBOT profile);
* ODK_RESOURCES_DIR/robot/plugins/... (ROBOT plugins).

This commit changes the way the ODK tools are installed so that all the
ODK stuff is cleanly organised in a single /odk directory as follows:

* /odk/bin: executable programs (may be links to somewhere else);
* /odk/tools: additional files that a tool may need to run (typically,
  the Jar archives for a Java tool);
* /odk/resources: the ODK_RESOURCES_DIR as explained above.

We install all programs as usual within this layout, then we run the
`odk install` command; the command will only install a tool if the tool
is not already available, so it will _not_ reinstall any tool that has
already been "manually" installed previously in the Dockerfile -- but
that will ensure that, if ODKCore starts needing a new tool, that tool
will always be included in the ODKLite image even if we do not update
the Dockerfile. This principle also allows us to override the versions
of the tools set forth by ODKCore.

We remove the context2csv.py and check-rdfxml.sh scripts, which are
replaced by subcommands of ODKCore's odk-helper command.