Re-add Jinjanator to ODKFull.
Jinjanator had been mistakenly removed from both ODKLite and ODKFull; it
does not belong to ODKLite but we expect its presence in ODKFull, so we
restore it there.

We also move the test for the presence of both Jinjanator and Owltools'
ontology-release-runner from the `test_odklite_programs` target to the
`test_odkfull_programs` target, since both those tools belong only to
ODKFull.