Add odk.py symlinks for compatibility.
The main `odk` script used to be located at `/tools/odk.py`, where it
could be invoked either as `/tools/odk.py` (full path, hardcoded) or as
`odk.py` (since `/tools` was in the PATH).

Now the script lives in `/usr/local/bin/odk`, and is expected to be
invoked simply with `odk`.

But existing calls to either `/tools/odk.py` or `odk.py` should still
work for compatibility. In particular, the `seed-via-docker.sh` script
must be usable with both pre-1.7 ODK images (where the script is under
`/tools/odk.py`) and post-1.7 images.

To ensure that, we add symlinks to the image so that `/tools/odk.py` and
`/usr/local/bin/odk.py` both point to `/usr/local/bin/odk`.