Use local odk-core submodule when building the images.
Now that we have the `odkcore` repository as a submodule under the
`core` directory, we can refer to that directory to install the
`odk-core` Python module when building the `odkbuild` image.

This requires:

* replacing the `odk-core` name by an explicit path reference;
* binding the `core` directory to `/build/core`, when building the
  `odkbuild` image, so that PIP can resolve the path reference.

Additionally, when computing the constraints, we must explicitly
*exclude* the `odk-core` package, because `pip freeze` records it as
something like:

```
odk-core @ file:/path/to/odk/core
```

instead of using a version number, and this will then cause `pip
install` to be unable to install `odk-core` without violating the
constraint.