Only run a subset of the configuration tests.
There should be no need, after building a ODK image, to run _all_ the
configuration tests from ODK Core. Those tests will have been run
already as part of the ODK Core test suite.

It is still valuable to run _some_ configuration/seeding tests, to be
sure that the ODK Core works just as fine when it is part of a ODK image
as when it is used in a standalone, native ODK environment, but for that
it should be enough to test a handful of configurations.