Remove redundant test.
There is no need to explicitly call the `test_programs_odkfull` target
as part of `make tests`, since that target is already called as part of
the generic `test-flavor` target.

So, `make tests` becomes merely a synonym of `make test-full`. It is not
used anywhere else in the project (the GitHub Actions workflow calls
`make test-full` directly), but it may still be useful to have such a
synonym (which we could redirect to another test target later on), so
we'll keep it.