Remove the JSON schema for the configuration object.
The auto-generated JSON schema for ODK configuration objects has not
been updated for nearly two years, and currently _cannot_ be updated
because of some issues with the `dataclasses-jsonschema` module (issues
that will never be fixed since that module is no longer maintained).

Out-of-date documentation is already bad enough, but arguably an
out-of-date _schema_ is even worse, because where there is a schema,
people will legitimately expect to be able to validate their
configuration files against the schema. Better not to provide any schema
than to provide an out-of-date, inaccurate one.